+++ libmscore/vibrato.cpp
@@ -10,17 +10,16 @@
 //  the file LICENCE.GPL
 //=============================================================================
 
-#include "vibrato.h"
-#include "style.h"
-#include "system.h"
 #include "measure.h"
-#include "xml.h"
-#include "utils.h"
-#include "sym.h"
 #include "score.h"
-#include "accidental.h"
 #include "segment.h"
 #include "staff.h"
+#include "style.h"
+#include "sym.h"
+#include "system.h"
+#include "utils.h"
+#include "vibrato.h"
+#include "xml.h"
 
 namespace Ms {
 
@@ -66,7 +65,7 @@ void VibratoSegment::symbolLine(SymId start, SymId fill)
       _symbols.push_back(start);
       qreal w1 = f->advance(start, mag);
       qreal w2 = f->advance(fill, mag);
-      int n    = lrint((w - w1) / w2);
+      int n    = (int)lrint((w - w1) / w2);
       for (int i = 0; i < n; ++i)
            _symbols.push_back(fill);
       QRectF r(f->bbox(_symbols, mag));
@@ -86,7 +85,7 @@ void VibratoSegment::symbolLine(SymId start, SymId fill, SymId end)
       qreal w1 = f->bbox(start, mag).width();
       qreal w2 = f->width(fill, mag);
       qreal w3 = f->width(end, mag);
-      int n    = lrint((w - w1 - w3) / w2);
+      int n    = (int)lrint((w - w1 - w3) / w2);
       for (int i = 0; i < n; ++i)
            _symbols.push_back(fill);
       _symbols.push_back(end);
